% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_n.R
\name{extract_norm_group_fim}
\alias{extract_norm_group_fim}
\title{Extract a normalized group FIM}
\usage{
extract_norm_group_fim(poped.db, ...)
}
\arguments{
\item{poped.db}{A PopED database}

\item{...}{Arguments passed to \code{\link{evaluate.fim}}}
}
\value{
A list of FIMs, one for each group in a design.
}
\description{
Extract an individual FIM for each group in a design
}
\keyword{internal}
