% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileparts.R
\name{fileparts}
\alias{fileparts}
\title{MATLAB fileparts function}
\usage{
fileparts(filename.with.path)
}
\arguments{
\item{filename.with.path}{A string of a filename with a path}
}
\value{
A list with the following components:
\item{pathname}{The path name}
\item{filename}{The file name}
\item{fileext}{The file extension}
}
\description{
Get the various parts of a file with path string.
}
\note{
This is a modified version of the same function in the matlab R-package.
}
\examples{

fileparts("ggg/ttt/lll.R")

}
\seealso{
Other MATLAB:
\code{\link[=cell]{cell()}},
\code{\link[=diag_matlab]{diag_matlab()}},
\code{\link[=feval]{feval()}},
\code{\link[=isempty]{isempty()}},
\code{\link[=ones]{ones()}},
\code{\link[=rand]{rand()}},
\code{\link[=randn]{randn()}},
\code{\link[=size]{size()}},
\code{\link[=tic]{tic()}},
\code{\link[=toc]{toc()}},
\code{\link[=zeros]{zeros()}}
}
\concept{MATLAB}
\keyword{internal}
