% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTruncatedNormal.R
\name{getTruncatedNormal}
\alias{getTruncatedNormal}
\title{Generate a random sample from a truncated normal distribution.}
\usage{
getTruncatedNormal(mean, variance)
}
\arguments{
\item{mean}{the mean of the normal distribution}

\item{variance}{The variance of the normal distribution}
}
\value{
A random sample from the specified truncated normal distribution
}
\description{
Generate a random sample from a truncated normal distribution.
}
\examples{

getTruncatedNormal(mean=3,variance=100)


}
\keyword{internal}
