% Generated by roxygen2 (4.0.1): do not edit by hand
\name{isfield}
\alias{isfield}
\title{Function written to match MATLAB's isfield function}
\usage{
isfield(obj, sub.obj.str)
}
\arguments{
\item{obj}{A list or dataframe}

\item{sub.obj.str}{A string giving the name of the sub-object you want to check for.}
}
\value{
Logical. True if the element exists.
}
\description{
Check if a list or dataframe has an element with a specific name.
}
\examples{

foo <- list("fff"=8,"ggg"=9)

isfield(foo,"fff")
isfield(foo,"lll")
}
\seealso{
Other MATLAB: \code{\link{cell}};
  \code{\link{diag_matlab}}; \code{\link{feval}};
  \code{\link{fileparts}}; \code{\link{isempty}};
  \code{\link{ones}}; \code{\link{randn}};
  \code{\link{randperm}}; \code{\link{rand}};
  \code{\link{size}}; \code{\link{tic}}; \code{\link{toc}};
  \code{\link{zeros}}
}

