% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_hilow_poped.R
\name{median_hilow_poped}
\alias{median_hilow_poped}
\title{Wrap summary functions from Hmisc and ggplot to work with stat_summary in ggplot}
\usage{
median_hilow_poped(x, ...)
}
\arguments{
\item{x}{A numeric vector}

\item{...}{Additional arguments passed to Hmisc's smedian.hilow function or 
ggplot2's median_hilow function, depending on your version of ggplot.}
}
\description{
Created for back compatibility with older versions of ggplot, and so that 
PopED does not have to load ggplot when started.
}
