% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ones.R
\name{ones}
\alias{ones}
\title{Create a matrix of ones}
\usage{
ones(dim1, dim2 = NULL)
}
\arguments{
\item{dim1}{The dimension of the matrix (if square) or the number of rows.}

\item{dim2}{The number of columns}
}
\value{
A matrix of ones
}
\description{
Create a matrix of ones of size (dim1 x dim2).
}
\examples{

ones(4)
ones(3,4)
}
\seealso{
Other MATLAB:
\code{\link[=cell]{cell()}},
\code{\link[=diag_matlab]{diag_matlab()}},
\code{\link[=feval]{feval()}},
\code{\link[=fileparts]{fileparts()}},
\code{\link[=isempty]{isempty()}},
\code{\link[=rand]{rand()}},
\code{\link[=randn]{randn()}},
\code{\link[=size]{size()}},
\code{\link[=tic]{tic()}},
\code{\link[=toc]{toc()}},
\code{\link[=zeros]{zeros()}}
}
\concept{MATLAB}
