% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_n.R
\name{optimize_n_rse}
\alias{optimize_n_rse}
\title{Optimize the number of subjects based on desired uncertainty of a parameter.}
\usage{
optimize_n_rse(
  poped.db,
  bpop_idx,
  need_rse,
  use_percent = TRUE,
  allowed_values = seq(poped.db$design$m, sum(poped.db$design$groupsize) * 5, by =
    poped.db$design$m)
)
}
\arguments{
\item{poped.db}{A PopED database.}

\item{bpop_idx}{The index number of the parameter, currently only bpop parameters are allowed.}

\item{need_rse}{The relative standard error (RSE) one would like to achieve (in percent, by default).}

\item{use_percent}{Should the RSE be represented as a percentage (T/F)?}

\item{allowed_values}{A vector of the allowed total number of subjects in the study.}
}
\value{

}
\description{
Optimize the number of subjects, 
based on the current design and the desired uncertainty of a single parameter
}
\examples{
tests/testthat/examples_fcn_doc/examples_optimize_n_rse.R
}
