\name{PopED-internal}
\alias{PopED-internal}
\alias{calc_k}
\alias{blockoptwrt}
\alias{blockother}
\alias{blockother_2}
\alias{cauchy_point}
\alias{d2fimdalpha2}
\alias{datestr_poped}
\alias{ceil}
\alias{compute_step}
\alias{dfimdalpha}
\alias{find.largest.index}
\alias{find_matlab}
\alias{fprintf}
\alias{fullfile}
\alias{get_cov_matrix_index}
\alias{get_discrete_val}
\alias{get_discrete_x}
\alias{get_fim_size}
\alias{get_step_size}
\alias{get_unfixed_params}
\alias{get_vector_str}
\alias{hessian_eta_complex}
\alias{hesskalpha2}
\alias{isinf}
\alias{inv}
\alias{kron_tmp}
\alias{limit_value}
\alias{matrix_all}
\alias{matrix_any}
\alias{reshape_matlab}
\alias{subspace_min}
\alias{write_iterationfile}
\alias{v}
\alias{writet}
\alias{write_matrix}
\alias{update_offdiag}
\alias{update_designinlist}
\alias{trace_matrix}
\alias{rem}
\alias{projected_gradient_norm}
\alias{print_xt}
\alias{print_params}
\alias{print_information}
\alias{own_normpdf}
\alias{numel}
\alias{new_ferror_file}
\alias{min_function}
\alias{log_prior_pdf}
\alias{linspace}
\alias{line_search_uc}
\alias{line_search}
\alias{isnan}
\alias{get_cv}
\alias{get_parnam}
\alias{graddetmfa}
\alias{graddetmfa_ext}
\alias{graddetmfxt}
\alias{graddetmfxt_ext}
\alias{gradff}
\alias{gradfg}
\alias{gradfg_occ}
\alias{gradlndetmfa}
\alias{gradlndetmfa_ext}
\alias{gradlndetmfxt}
\alias{gradlndetmfxt_ext}
\alias{gradofv_a}
\alias{gradofv_xt}
\alias{gradtrmfa}
\alias{gradtrmfxt}
\alias{graph_det}
\alias{grouped_rand}
\alias{grouped_rand_a}
\alias{ind_estimates}
\alias{ind_likelihood}
\alias{m1}
\alias{m2}
\alias{m3}
\alias{m4}
\alias{mf_all}
\alias{sg_search}

\title{Internal PopED Functions}
\description{
  Internal PopED functions
}
\details{
  These functions will typically not be called by the user (or in some cases are just
  waiting for proper documentation to be written).
}
\keyword{ internal }