% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popED.R
\name{popED}
\alias{popED}
\title{Population Distributed Equally}
\usage{
popED(mastergrid, px_area, pop, cores = 4, output_dir=tempdir(), 
prefix="A1", nt=2, check_result=TRUE, verbose = TRUE)
}
\arguments{
\item{mastergrid}{Raster tat contains the unique area IDs as their value.}

\item{px_area}{Raster containing the pixel area.}

\item{pop}{Character vector containing the name of the file from which the
unique area ID and corresponding population values are to be read
from. The file should contain two columns comma-separated with the
value of administrative ID and population without columns names.
If it does not contain an absolute path, the file name is relative to
the current working directory.}

\item{cores}{Integer vector containing an integer. Indicates the number of
cores to use in parallel when executing the function. If set to 0
\code{(max_number_of_cores - 1)}  will be used based on as many
processors as the hardware and RAM allow. Default is \code{cores} = 0.}

\item{output_dir}{Character vector containing the path to the directory for
writing output files. Default is the temp directory.}

\item{prefix}{Character to be added to the output results. Default is "A1"}

\item{nt}{Numeric. Increase number of blocks sugesting for
processing raster file. Default is \code{nt} = 2}

\item{check_result}{Logical vector indicating whether the results will be
compared with input data. Default is \code{check_result} = TRUE.}

\item{verbose}{Logical vector indicating whether to print
intermediate output from the function to the console, which might be
helpful for model debugging. Default is \code{verbose} = TRUE.}
}
\value{
Raster* object of gridded population equally distributed.
}
\description{
Population Distributed Equally
}
\details{
This function produces gridded population equally distributed
}
\examples{
\dontrun{

library("popED")

res <- popED(mastergrid="input_mastergrid.tif", 
             px_area="input_px_area.tif", 
             pop="pop_table.csv",
             cores=4,
             output_dir="/user/output" 
             ) 
 
# Plot populataion raster 
plot(res) 

}
}
\author{
Maksym Bondarenko \href{mailto:mb4@soton.ac.uk}{mb4@soton.ac.uk},
David Kerr \href{mailto:dk2n16@soton.ac.uk}{dk2n16@soton.ac.uk},
Alessandro Sorichetta \href{mailto:as1v13@soton.ac.uk}{as1v13@soton.ac.uk} and
Tom McKeen \href{mailto:t.l.mckeen@soton.ac.uk}{t.l.mckeen@soton.ac.uk}
}
