% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.R
\name{rand}
\alias{rand}
\title{Function written to match MATLAB's rand function}
\usage{
rand(dim1, dim2 = NULL)
}
\arguments{
\item{dim1}{The dimension of the matrix (if square), otherwise the number of rows.}

\item{dim2}{The number of columns, if different from the number of rows.}
}
\value{
Matrix of random generated samples.
}
\description{
Generate random samples from a uniform distribution [0,1] and return in matrix form
}
\examples{

rand(2,3)

rand(5)
}
\seealso{
Other MATLAB:
\code{\link[=cell]{cell()}},
\code{\link[=diag_matlab]{diag_matlab()}},
\code{\link[=feval]{feval()}},
\code{\link[=fileparts]{fileparts()}},
\code{\link[=isempty]{isempty()}},
\code{\link[=ones]{ones()}},
\code{\link[=randn]{randn()}},
\code{\link[=size]{size()}},
\code{\link[=tic]{tic()}},
\code{\link[=toc]{toc()}},
\code{\link[=zeros]{zeros()}}
}
\concept{MATLAB}
\keyword{internal}
