% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randn.R
\name{randn}
\alias{randn}
\title{Function written to match MATLAB's randn function}
\usage{
randn(dim1, dim2 = NULL)
}
\arguments{
\item{dim1}{The dimension of the matrix (if square), otherwise the number of rows.}

\item{dim2}{The number of columns, if different from the number of rows.}
}
\value{
Matrix of random generated samples.
}
\description{
Generate random samples from a standardized normal distribution and return in matrix form.
}
\examples{

randn(2,3)

randn(5)
}
\seealso{
Other MATLAB:
\code{\link[=cell]{cell()}},
\code{\link[=diag_matlab]{diag_matlab()}},
\code{\link[=feval]{feval()}},
\code{\link[=fileparts]{fileparts()}},
\code{\link[=isempty]{isempty()}},
\code{\link[=ones]{ones()}},
\code{\link[=rand]{rand()}},
\code{\link[=size]{size()}},
\code{\link[=tic]{tic()}},
\code{\link[=toc]{toc()}},
\code{\link[=zeros]{zeros()}}
}
\concept{MATLAB}
\keyword{internal}
