% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{tryCatch.W.E}
\alias{tryCatch.W.E}
\title{tryCatch both warnings (with value) and errors}
\usage{
tryCatch.W.E(expr)
}
\arguments{
\item{expr}{an \R expression to evaluate}
}
\value{
a list with 'value' and 'warning', where
  'value' may be an error caught.
}
\description{
Catch *and* save both errors and warnings, and in the case of
a warning, also keep the computed result.
}
\author{
Martin Maechler, The R Core Team
}
\keyword{internal}
