% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeros.R
\name{zeros}
\alias{zeros}
\title{Create a matrix of zeros.}
\usage{
zeros(dim1, dim2 = NULL)
}
\arguments{
\item{dim1}{The dimension of the matrix (if square) or the number of rows.}

\item{dim2}{The number of columns}
}
\value{
A matrix of zeros.
}
\description{
Create a matrix of zeros of size (dim1 x dim2).
}
\examples{

zeros(3)
zeros(0,3)
zeros(4,7)
zeros(1,4)
}
\seealso{
Other MATLAB:
\code{\link[=cell]{cell()}},
\code{\link[=diag_matlab]{diag_matlab()}},
\code{\link[=feval]{feval()}},
\code{\link[=fileparts]{fileparts()}},
\code{\link[=isempty]{isempty()}},
\code{\link[=ones]{ones()}},
\code{\link[=rand]{rand()}},
\code{\link[=randn]{randn()}},
\code{\link[=size]{size()}},
\code{\link[=tic]{tic()}},
\code{\link[=toc]{toc()}}
}
\concept{MATLAB}
