\name{blockopt}
\alias{blockopt}
\title{Summarize your optimization settings for optimization routines}
\usage{
blockopt(fn, poped.db)
}
\arguments{
  \item{poped.db}{A PopED database.}

  \item{fn}{The file handle to write to.}
}
\description{
Create some output to the screen and a text file that
summarizes the optimization settings you will use to
optimize.
}
\seealso{
Other Helper: \code{\link{blockexp}};
\code{\link{blockfinal_2}}; \code{\link{blockfinal}};
\code{\link{blockheader_2}}; \code{\link{blockheader}};
\code{\link{blockopt_2}}
}

