\name{cell}
\alias{cell}
\title{MATLAB cell function}
\usage{
cell(...)
}
\arguments{
  \item{...}{Dimensions for the cell array.}
}
\value{
A list of empty lists.
}
\description{
Create a cell array as in MATLAB.
}
\note{
This is a modified version of the same function in
\code{\link[matlab]{cell}}
}
\seealso{
Other MATLAB: \code{\link{PopED}},
\code{\link{PopED-package}}, \code{\link{poped}};
\code{\link{diag_matlab}}; \code{\link{feval}};
\code{\link{fileparts}}; \code{\link{isempty}};
\code{\link{isfield}}; \code{\link{ones}};
\code{\link{randn}}; \code{\link{randperm}};
\code{\link{rand}}; \code{\link{size}}; \code{\link{tic}};
\code{\link{toc}}; \code{\link{zeros}}
}

