% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_summary.R
\name{design_summary}
\alias{design_summary}
\title{Display a summary of output from poped_db}
\usage{
design_summary(poped_db, file = "", ...)
}
\arguments{
\item{poped_db}{An object returned from \code{\link{create.poped.database}} to summarize.}

\item{file}{A file handle to write to.  Default is to the R console.}

\item{...}{Additional arguments. Passed to \code{\link{blockfinal}}.}
}
\description{
Display a summary of output from poped_db
}
