% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feval.R
\name{feval}
\alias{feval}
\title{MATLAB feval function}
\usage{
feval(file.name, ...)
}
\arguments{
\item{file.name}{A function or a string that is the name of a function.}

\item{...}{Arguments for the function.  Multiple arguments separated by a comma.}
}
\value{
Output from the defined function.
}
\description{
This is just a wrapper for the \code{\link{do.call}} function to behave like the feval function in MATLAB.
}
\examples{

feval("sin",pi/2)

}
\seealso{
Other MATLAB: \code{\link{cell}},
  \code{\link{diag_matlab}}, \code{\link{fileparts}},
  \code{\link{isempty}}, \code{\link{isfield}},
  \code{\link{ones}}, \code{\link{randn}},
  \code{\link{randperm}}, \code{\link{rand}},
  \code{\link{size}}, \code{\link{tic}}, \code{\link{toc}},
  \code{\link{zeros}}
}
\keyword{internal}

