% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfulld.R
\name{getfulld}
\alias{getfulld}
\title{Create a full D (between subject variability) matrix given a vector of variances and covariances.}
\usage{
getfulld(variance_vector, covariance_vector = NULL)
}
\arguments{
\item{variance_vector}{The vector of the variances.}

\item{covariance_vector}{A vector of the covariances. Writen in column major 
order for the lower triangular matrix.}
}
\value{
The full matrix of variances for the between subject variances
}
\description{
Create a full D (between subject variability) matrix given a vector of variances and covariances.
}
\examples{

getfulld(c(1,2,3))

getfulld(c(1,2,3),c(7,6,5))

}
\keyword{internal}

