% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv.R
\name{inv}
\alias{inv}
\title{Compute the inverse of a matrix}
\usage{
inv(mat)
}
\arguments{
\item{mat}{A matrix}
}
\value{
The inverse matrix
}
\description{
Function computes the inverse of a matrix.
}
\keyword{internal}

