% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_prior_pdf.R
\name{log_prior_pdf}
\alias{log_prior_pdf}
\title{Compute the natural log of the PDF for the parameters in an E-family design}
\usage{
log_prior_pdf(alpha, bpopdescr, ddescr, return_gradient = F,
  return_hessian = F)
}
\arguments{
\item{alpha}{A parameter vector.}

\item{bpopdescr}{Matrix defining the fixed effects, per row (row number =
parameter_number) we should have: \itemize{ \item column 1 the type of the
distribution for E-family designs (0 = Fixed, 1 = Normal, 2 = Uniform, 3 =
User Defined Distribution, 4 = lognormal and 5 = truncated normal) \item
column 2  defines the mean. \item column 3 defines the variance of the
distribution (or length of uniform distribution). }}

\item{ddescr}{Matrix defining the diagonals of the IIV (same logic as for
the \code{bpopdescr}).}

\item{return_gradient}{Should the gradient be returned.}

\item{return_hessian}{Should the hessian be returned?}
}
\description{
Compute the natural log of the PDF for the parameters in an E-family design
}
\keyword{internal}
