% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/optim_ARS.R
\name{optim_ARS}
\alias{optim_ARS}
\title{Optimization Using Adaptive Random Search.}
\usage{
optim_ARS(par, fn, lower, upper, allowed_values = NULL, loc_fac = 4,
  iter = 400, iter_adapt = 50, max_run = 200, iter_chunk = NULL,
  trace = TRUE, trace_iter = 5, par_grouping = NULL,
  new_par_max_it = 200, minimize = F, parallel = F,
  parallel_type = NULL, num_cores = NULL, seed = round(runif(1, 0,
  1e+07)), ...)
}
\arguments{
\item{loc_fac}{Locality factor}
}
\description{
Optimize an objective function using an adaptive random search algorithm.
The function works for both discrete and continuous optimization parameters
and allows for a number of constraints (box-constraints, sets of allowed values,
constraints between parameters to be optimized).
}
\references{
\enumerate{
\item M. Foracchia, A.C. Hooker, P. Vicini and A. Ruggeri, "PopED, a software fir optimal
experimental design in population kinetics", Computer Methods and Programs in Biomedicine, 74, 2004.
\item J. Nyberg, S. Ueckert, E.A. Stroemberg, S. Hennig, M.O. Karlsson and A.C. Hooker, "PopED: An extended,
parallelized, nonlinear mixed effects models optimal design tool",
Computer Methods and Programs in Biomedicine, 108, 2012.
}
}
\seealso{
Other Optimize: \code{\link{Doptim}};
  \code{\link{LEDoptim}}; \code{\link{RS_opt_gen}};
  \code{\link{RS_opt}}; \code{\link{a_line_search}};
  \code{\link{bfgsb_min}}; \code{\link{calc_autofocus}};
  \code{\link{calc_ofv_and_grad}}; \code{\link{mfea}};
  \code{\link{optim_LS}}; \code{\link{poped_optimize}};
  \code{\link{poped_optim}}
}

