\name{PopED-package}
\alias{PopED-package}
\alias{PopED}
\docType{package}
\title{
PopED - \bold{Pop}ulation (and individual) optimal \bold{E}xperimental \bold{D}esign. 
}
\description{
PopED computes optimal experimental designs for both 
  population  and individual studies based on nonlinear mixed-effect models.  
  Often this is based on a computation of the Fisher Information Matrix (FIM). 
}
\details{
\tabular{ll}{
Package: \tab PopED\cr
Version: \tab 0.0.1\cr
Date: \tab 2014-01-04\cr
License: \tab LGPL >= 3\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Andrew Hooker <Andrew.Hooker@farmbio.uu.se>
}
\references{
\enumerate{
\item M. Foracchia, A.C. Hooker, P. Vicini and A. Ruggeri, "PopED, a software fir optimal 
experimental design in population kinetics", Computer Methods and Programs in Biomedicine, 74, 2004.
\item J. Nyberg, S. Ueckert, E.A. Stroemberg, S. Hennig, M.O. Karlsson and A.C. Hooker, "PopED: An extended, 
parallelized, nonlinear mixed effects models optimal design tool",  
Computer Methods and Programs in Biomedicine, 108, 2012.
\item \href{http://poped.sourceforge.net}{poped.sf.net}
}
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
%~~ directory ~~
\keyword{ package }
%\seealso{
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
