\docType{package}
\name{PopED-package}
\alias{PopED}
\alias{PopED-package}
\alias{poped}
\title{PopED - \bold{Pop}ulation (and individual) optimal \bold{E}xperimental \bold{D}esign.}
\description{
PopED computes optimal experimental designs for both
population and individual studies based on nonlinear
mixed-effect models. Often this is based on a computation
of the Fisher Information Matrix (FIM).
}
\references{
\enumerate{ \item J. Nyberg, S. Ueckert, E.A. Stroemberg,
S. Hennig, M.O. Karlsson and A.C. Hooker, "PopED: An
extended, parallelized, nonlinear mixed effects models
optimal design tool", Computer Methods and Programs in
Biomedicine, 108, 2012. \item M. Foracchia, A.C. Hooker, P.
Vicini and A. Ruggeri, "PopED, a software for optimal
experimental design in population kinetics", Computer
Methods and Programs in Biomedicine, 74, 2004. \item
\href{http://poped.sourceforge.net}{poped.sf.net} \item
\url{https://github.com/andrewhooker/PopED.git} }
}
\seealso{
Other E-family: \code{\link{ed_laplace_ofv}};
\code{\link{ed_mftot}}; \code{\link{evaluate.e.ofv.fim}}

Other FIM: \code{\link{LinMatrixH}};
\code{\link{LinMatrixLH}}; \code{\link{LinMatrixL_occ}};
\code{\link{ed_laplace_ofv}}; \code{\link{ed_mftot}};
\code{\link{evaluate.e.ofv.fim}};
\code{\link{evaluate.fim}}; \code{\link{gradf_eps}};
\code{\link{mf3}}; \code{\link{mf5}}; \code{\link{mf6}};
\code{\link{mf7}}; \code{\link{mf8}}; \code{\link{mftot0}};
\code{\link{mftot1}}; \code{\link{mftot2}};
\code{\link{mftot3}}; \code{\link{mftot4}};
\code{\link{mftot5}}; \code{\link{mftot6}};
\code{\link{mftot7}}; \code{\link{mftot}};
\code{\link{mf}}; \code{\link{ofv_criterion}};
\code{\link{ofv_fim}}

Other Graphics: \code{\link{plot_efficiency_of_windows}};
\code{\link{plot_model_prediction}}

Other Helper: \code{\link{blockexp}};
\code{\link{blockfinal_2}}; \code{\link{blockfinal}};
\code{\link{blockheader_2}}; \code{\link{blockheader}};
\code{\link{blockopt_2}}; \code{\link{blockopt}}

Other MATLAB: \code{\link{cell}};
\code{\link{diag_matlab}}; \code{\link{feval}};
\code{\link{fileparts}}; \code{\link{isempty}};
\code{\link{isfield}}; \code{\link{ones}};
\code{\link{randn}}; \code{\link{randperm}};
\code{\link{rand}}; \code{\link{size}}; \code{\link{tic}};
\code{\link{toc}}; \code{\link{zeros}}

Other Optimize: \code{\link{Doptim}};
\code{\link{a_line_search}}; \code{\link{bfgsb_min}};
\code{\link{calc_autofocus}};
\code{\link{calc_ofv_and_grad}}; \code{\link{mfea}};
\code{\link{poped_optimize}}

Other Simulation: \code{\link{model_prediction}};
\code{\link{plot_efficiency_of_windows}};
\code{\link{plot_model_prediction}}

Other evaluate_design: \code{\link{evaluate.fim}};
\code{\link{get_rse}}; \code{\link{model_prediction}};
\code{\link{plot_efficiency_of_windows}};
\code{\link{plot_model_prediction}}

Other matrix_manipulation: \code{\link{diag_matlab}};
\code{\link{test_for_max}}; \code{\link{test_for_min}};
\code{\link{test_for_zeros}}

Other poped_input: \code{\link{convert_variables}};
\code{\link{create.poped.database}};
\code{\link{downsizing_general_design}};
\code{\link{poped.choose}}
}

