\name{size}
\alias{size}
\title{Function written to match MATLAB's size function}
\usage{
size(obj, dimension.index = NULL)
}
\arguments{
  \item{obj}{An object you want to know the various
  dimensions of.  Typically a matrix.}

  \item{dimension.index}{Which dimension you are interested
  in.}
}
\value{
The dimensions of the object or specific dimension you are
interested in.
}
\description{
Function written to match MATLAB's size function
}
\seealso{
Other MATLAB: \code{\link{PopED}},
\code{\link{PopED-package}}, \code{\link{poped}};
\code{\link{cell}}; \code{\link{diag_matlab}};
\code{\link{feval}}; \code{\link{fileparts}};
\code{\link{isempty}}; \code{\link{isfield}};
\code{\link{ones}}; \code{\link{randn}};
\code{\link{randperm}}; \code{\link{rand}};
\code{\link{tic}}; \code{\link{toc}}; \code{\link{zeros}}
}

