% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/test_for_max.R
\name{test_for_max}
\alias{test_for_max}
\title{Test if any matrix element is above a max value.}
\usage{
test_for_max(mat, max_mat)
}
\arguments{
\item{mat}{A matrix.}

\item{max_mat}{A matrix the same size as mat with the  maximum allowed value of that element.}
}
\value{
A matrix
}
\description{
Function tests is any matrix element is above a maximum value.  For those elements
the function sets those values to the maximum value.
}
\examples{

test_for_max(cbind(2,3,4,5,6),cbind(4,4,4,4,4))

test_for_max(ones(6)*45,ones(6)*40)



}
\seealso{
Other matrix_manipulation: \code{\link{diag_matlab}};
  \code{\link{test_for_min}}; \code{\link{test_for_zeros}}
}

