% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_for_min.R
\name{test_for_min}
\alias{test_for_min}
\title{Test if any matrix element is above a minimum value.}
\usage{
test_for_min(mat, min_mat)
}
\arguments{
\item{mat}{A matrix.}

\item{min_mat}{A matrix the same size as mat with the  minimum allowed value of that element.}
}
\value{
A matrix
}
\description{
Function tests is any matrix element is above a minimum value.  For those elements 
the function sets those values to the minimum value.
}
\examples{

test_for_min(cbind(2,3,4,5,6),cbind(4,4,4,4,4))

test_for_min(ones(6)*45,ones(6)*40)




}
\seealso{
Other matrix_manipulation: \code{\link{diag_matlab}},
  \code{\link{test_for_max}}, \code{\link{test_for_zeros}}
}
\keyword{internal}

