% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_for_zeros.R
\name{test_for_zeros}
\alias{test_for_zeros}
\title{Test if any matrix element is zero.}
\usage{
test_for_zeros(mat, ourzero)
}
\arguments{
\item{mat}{A matrix.}

\item{ourzero}{A matrix the same size as mat with the  value that zero should be reassigned to.}
}
\value{
A matrix
}
\description{
Function tests is any matrix element is zero.  For those elements 
the function sets those values to the minimum value allowed (not zero).
This is used to avoid numerical problems in the FIM calculation.
}
\examples{

test_for_zeros(cbind(2,3,0,5,6),1e-18)

test_for_zeros(zeros(6),1e-7)

test_for_zeros(cbind(1,2,3,NA,0),1e-5)



}
\seealso{
Other matrix_manipulation: \code{\link{diag_matlab}},
  \code{\link{test_for_max}}, \code{\link{test_for_min}}
}
\keyword{internal}

