% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toc.R
\name{toc}
\alias{toc}
\title{Timer function (as in MATLAB)}
\usage{
toc(echo = TRUE, name = ".poped_savedTime")
}
\arguments{
\item{echo}{Print time to screen?}

\item{name}{The saved name of the time object.}
}
\description{
Function to stop a timer.  Start with tic().
}
\note{
This is a modified version of the same function in the matlab R-package.
}
\examples{

tic()
toc()

tic(name="foo")
toc()
tic()
toc()
toc()
tic()
toc(name="foo")
}
\seealso{
Other MATLAB: 
\code{\link{cell}()},
\code{\link{diag_matlab}()},
\code{\link{feval}()},
\code{\link{fileparts}()},
\code{\link{isempty}()},
\code{\link{ones}()},
\code{\link{randn}()},
\code{\link{rand}()},
\code{\link{size}()},
\code{\link{tic}()},
\code{\link{zeros}()}
}
\concept{MATLAB}
