% Generated by roxygen2 (4.0.1): do not edit by hand
\name{toc}
\alias{toc}
\title{Timer function (as in MATLAB)}
\usage{
toc(echo = TRUE, name = ".poped_savedTime")
}
\arguments{
\item{echo}{Print time to screen?}

\item{name}{The saved name of the time object.}
}
\description{
Funtion to stop a timer.  Start with tic().
}
\note{
This is a modified version of the same function in the matlab package \code{\link[matlab]{toc}}
}
\examples{

tic()
toc()

tic(name="foo")
toc()
tic()
toc()
toc()
tic()
toc(name="foo")
}
\seealso{
Other MATLAB: \code{\link{cell}};
  \code{\link{diag_matlab}}; \code{\link{feval}};
  \code{\link{fileparts}}; \code{\link{isempty}};
  \code{\link{isfield}}; \code{\link{ones}};
  \code{\link{randn}}; \code{\link{randperm}};
  \code{\link{rand}}; \code{\link{size}};
  \code{\link{tic}}; \code{\link{zeros}}
}

