\name{zeros}
\alias{zeros}
\title{Function written to match MATLAB's zeros function}
\usage{
zeros(dim1, dim2 = NULL)
}
\arguments{
  \item{dim1}{The dimension of the matrix (if square) or
  the number of rows.}

  \item{dim2}{The number of columns}
}
\value{
A matrix of zeros.
}
\description{
Creates a matrix of zeros.
}
\seealso{
Other MATLAB: \code{\link{cell}};
\code{\link{diag_matlab}}; \code{\link{feval}};
\code{\link{fileparts}}; \code{\link{isempty}};
\code{\link{isfield}}; \code{\link{randn}};
\code{\link{randperm}}; \code{\link{rand}};
\code{\link{size}}
}

