% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeros.R
\name{zeros}
\alias{zeros}
\title{Creates a matrix of zeros.}
\usage{
zeros(dim1, dim2 = NULL)
}
\arguments{
\item{dim1}{The dimension of the matrix (if square) or the number of rows.}

\item{dim2}{The number of columns}
}
\value{
A matrix of zeros.
}
\description{
Function creates a matrix of zeros of size (dim1 x dim2). Written to match MATLAB's \code{zeros} function.
}
\examples{

zeros(3)
zeros(0,3)
zeros(4,7)
zeros(1,4)
}
\seealso{
Other MATLAB: \code{\link{cell}},
  \code{\link{diag_matlab}}, \code{\link{feval}},
  \code{\link{fileparts}}, \code{\link{isempty}},
  \code{\link{isfield}}, \code{\link{ones}},
  \code{\link{randn}}, \code{\link{randperm}},
  \code{\link{rand}}, \code{\link{size}},
  \code{\link{tic}}, \code{\link{toc}}
}

